@@ -4,7 +4,7 @@ description: Demonstrates how to use the connectivity plugin.
 dependencies:
   flutter:
     sdk: flutter
-  connectivity: any
+  connectivity_platform_interface: ^2.0.0
   connectivity_macos:
     # When depending on this package from a real application you should use:
     #   connectivity_macos: ^x.y.z
@@ -18,11 +18,11 @@ dev_dependencies:
     sdk: flutter
   integration_test:
     path: ../../../integration_test
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
   flutter: ">=1.10.0"
