@@ -1,21 +1,27 @@
 name: file_selector
 description: Flutter plugin for opening and saving files.
 homepage: https://github.com/flutter/plugins/tree/master/packages/file_selector/file_selector
-version: 0.7.0+2
+version: 0.8.1
+
+flutter:
+  plugin:
+    platforms:
+      web:
+        default_package: file_selector_web
 
 dependencies:
   flutter:
     sdk: flutter
-  file_selector_platform_interface: ^1.0.0
+  file_selector_platform_interface: ^2.0.0
+  file_selector_web: ^0.8.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  test: ^1.3.0
-  mockito: ^4.1.1
-  plugin_platform_interface: ^1.0.0
-  pedantic: ^1.8.0
+  test: ^1.16.3
+  plugin_platform_interface: ">=1.0.0 <3.0.0"
+  pedantic: ^1.10.0
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.5"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
+  flutter: ">=1.20.0"
