@@ -6,23 +6,23 @@
 
 name: extension_google_sign_in_as_googleapis_auth
 description: A bridge package between google_sign_in and googleapis_auth, to create Authenticated Clients from google_sign_in user credentials.
-version: 1.0.4
+version: 2.0.0
 homepage: https://github.com/flutter/plugins/google_sign_in/extension_google_sign_in_as_googleapis_auth
 
 dependencies:
   flutter:
     sdk: flutter
-  google_sign_in: ^4.4.1
-  googleapis_auth: ^0.2.11+1
-  meta: ^1.1.8
-  http: ^0.12.1
+  google_sign_in: ^5.0.0
+  googleapis_auth: ^1.0.0
+  meta: ^1.3.0
+  http: ^0.13.0
 
 dev_dependencies:
-  mockito: ^4.1.1
-  pedantic: ^1.9.0
+  pedantic: ^1.10.0
+  test: ^1.16.3
   flutter_test:
     sdk: flutter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
   flutter: ">=1.12.13+hotfix.4"
