@@ -2,13 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// TODO(cyanglaz): Remove once Mockito is migrated to null safety.
-// @dart = 2.9
 import 'package:connectivity/connectivity.dart';
 import 'package:connectivity_platform_interface/connectivity_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-import 'package:mockito/mockito.dart';
+import 'package:test/fake.dart';
 
 const ConnectivityResult kCheckConnectivityResult = ConnectivityResult.wifi;
 const LocationAuthorizationStatus kRequestLocationResult =
@@ -18,8 +16,8 @@ const LocationAuthorizationStatus kGetLocationResult =
 
 void main() {
   group('Connectivity', () {
-    Connectivity connectivity;
-    MockConnectivityPlatform fakePlatform;
+    late Connectivity connectivity;
+    late MockConnectivityPlatform fakePlatform;
     setUp(() async {
       fakePlatform = MockConnectivityPlatform();
       ConnectivityPlatform.instance = fakePlatform;
@@ -33,7 +31,7 @@ void main() {
   });
 }
 
-class MockConnectivityPlatform extends Mock
+class MockConnectivityPlatform extends Fake
     with MockPlatformInterfaceMixin
     implements ConnectivityPlatform {
   Future<ConnectivityResult> checkConnectivity() async {
