@@ -25,7 +25,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
@@ -57,7 +57,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget _infoTile(String title, String subtitle) {
     return ListTile(
       title: Text(title),
-      subtitle: Text(subtitle ?? 'Not set'),
+      subtitle: Text(subtitle.isNotEmpty ? subtitle : 'Not set'),
     );
   }
 
